Add two new requirement files
The python packages listed in these two files are what originally
in the requirement-gdalX.X.txt files.
I'm putting them in their own file so as to not lose the original
contents.
It will be up to the next person who actually has need of them, and will
use them to actually modify the docker image and the make file to
consume these: Since it appears these are actually unused at this time.